+++ src/index.ts
@@ -18,10 +18,11 @@ function gbm(config: BuildConfig) {
 
 namespace gbm {
   export function loadBuilders(config:BuildConfig) { bm.loadBuilders(config); }
-  export function parallel(...args: BuildSet[]) { return new GBuildSet(...args); };
-  export function series(...args:BuildSet[]) { return [...args]; };
+  export function parallel(...args: BuildSet[]) { return new GBuildSet(...args); }
+  export function series(...args:BuildSet[]) { return [...args]; }
   export let GBuilder = GBuilderClass;
   export let GPlugin = GPluginClass;
+  export let utils = require('./utils/utils');
 }
 
 registerPropertiesFromFiles(gbm, upath.join(__dirname, './plugins/*.js'));