+++ src/utils/utils.ts
@@ -98,3 +98,5 @@ export let wait = (msec: number) => new Promise(res => setTimeout(res, msec));
 //     errorCount = 0;
 //   }
 // }
+
+export * from './process';